chore: extract inline node code to separate wrappers
